fixes #1078 close should not assert if it fails
There are circumstances where the file descriptor can be already
closed, or can fail for EIO or other errors which can be wildly
platform specific.  Crashing the program here isn't helping.  We're
doing all we can by closing the file descriptor and the callers of
this function assuredly are not going to use the file descriptor again.